+++ src/tech/v3/dataset/column.clj
@@ -1,13 +1,11 @@
 (ns tech.v3.dataset.column
   (:require [tech.v3.protocols.column :as col-proto]
             [tech.v3.dataset.impl.column :as col-impl]
-            [tech.v3.dataset.impl.column-data-process :as column-data-process]
             [tech.v3.dataset.impl.column-index-structure :as col-index-structure]
             [tech.v3.dataset.string-table :as str-table]
             [tech.v3.dataset.io.column-parsers :as column-parsers]
             [tech.v3.datatype :as dtype]
-            [tech.v3.datatype.protocols :as dtype-proto]
-            [tech.v3.parallel.for :as pfor])
+            [tech.v3.datatype.protocols :as dtype-proto])
   (:import [java.util List]
            [tech.v3.dataset.impl.column Column]
            [org.roaringbitmap RoaringBitmap]))
@@ -161,14 +159,14 @@ Implementations should check their metadata before doing calculations."
       (catch Throwable e
         nil))))
 
-
+;; TODO - match inference expectations
 (defn parse-column
   "parse a text or a str column, returning a new column with the same name but with
   a different datatype.  This method is single-threaded.
 
   parser-fn-or-kwd is nil by default and can the keyword :relaxed?  or a function that
-  must return one of parsed-value, :tech.ml.dataset.parse/missing in which case a
-  missing value will be added or :tech.ml.dataset.parse/parse-failure in which case the
+  must return one of parsed-value, :tech.v3.dataset/missing in which case a
+  missing value will be added or :tech.v3.dataset/parse-failure in which case the
   a missing index will be added and the string value will be recorded in the metadata's
   :unparsed-data, :unparsed-indexes entries.
 
@@ -183,10 +181,8 @@ Implementations should check their metadata before doing calculations."
          n-elems (dtype/ecount col-reader)]
      (dotimes [iter n-elems]
        (column-parsers/add-value! col-parser iter (col-reader iter)))
-
-     (let [{:keys [data missing metadata]}
-           (column-parsers/finalize! col-parser n-elems)]
-       (new-column colname data metadata missing))))
+     (new-column (assoc (column-parsers/finalize! col-parser n-elems)
+                        :tech.v3.dataset/name colname))))
   ([datatype col]
    (parse-column datatype col nil)))
 
@@ -195,20 +191,13 @@ Implementations should check their metadata before doing calculations."
   "Create a new column.  Data will scanned for missing values
   unless the full 4-argument pathway is used."
   ([name data]
-   (let [{coldata :data
-          scanned-missing :missing}
-         (column-data-process/scan-data-for-missing data)]
-     (new-column name coldata nil scanned-missing)))
+   (col-impl/new-column name data))
   ([name data metadata]
-   (let [{coldata :data
-          scanned-missing :missing}
-         (column-data-process/scan-data-for-missing data)]
-     (new-column name coldata metadata scanned-missing)))
+   (col-impl/new-column name data metadata))
   ([name data metadata missing]
-   (let [data (if-not (dtype/as-buffer data)
-                (:data (column-data-process/scan-data-for-missing data))
-                data)]
-     (col-impl/new-column name data metadata missing))))
+   (col-impl/new-column name data metadata missing))
+  ([data-or-data-map]
+   (col-impl/new-column data-or-data-map)))
 
 
 (defn extend-column-with-empty
@@ -230,11 +219,43 @@ Implementations should check their metadata before doing calculations."
 
 
 (defn column-map
-  "Map a scalar function across one or more columns.  New missing set becomes
-  the union of all the other missing sets.  Column is named :_unnamed.
-  This is the missing-set aware version of tech.v3.datatype/emap."
+  "Map a scalar function across one or more columns.
+  This is the semi-missing-set aware version of tech.v3.datatype/emap.
+
+  If res-dtype is nil then the result is scanned to infer datatype and
+  missing set.  res-dtype may also be a map of options:
+
+  Options:
+
+  * `:datatype` - Set the dataype of the result column.  If not given result is scanned
+  to infer result datatype and missing set.
+  * `:missing-fn` - if given, columns are first passed to missing-fn as a sequence and
+  this dictates the missing set.  Else the missing set is by scanning the results
+  during the inference process.  See `tech.v3.dataset.column/union-missing-sets` and
+  `tech.v3.dataset.column/intersect-missing-sets` for example functions to pass in
+  here."
   [map-fn res-dtype & args]
-  (col-impl/new-column :_unnamed
-                       (apply dtype/emap map-fn res-dtype args)
-                       nil
-                       (reduce dtype-proto/set-or (map col-proto/missing args))))
+  (let [res-opt-map (if (keyword? res-dtype)
+                      {:datatype res-dtype}
+                      (or res-dtype {}))
+        ;;object readers get scanned to infer datatype
+        res-dtype (res-opt-map :datatype :object)
+        missing-fn (res-opt-map :missing-fn)]
+    (col-impl/new-column :_unnamed
+                         (apply dtype/emap map-fn res-dtype args)
+                         nil
+                         (if missing-fn
+                           (missing-fn args)
+                           nil))))
+
+
+(defn union-missing-sets
+  "Union the missing sets of the columns returning a roaring bitmap"
+  ^RoaringBitmap [col-seq]
+  (reduce dtype-proto/set-or (map col-proto/missing col-seq)))
+
+
+(defn intersect-missing-sets
+  "Intersect the missing sets of the columns returning a roaring bitmap"
+  ^RoaringBitmap [col-seq]
+  (reduce dtype-proto/set-and (map col-proto/missing col-seq)))